+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the Neo N3 Visual DevTracker extension will be documented
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## Unreleased
+
+### Changed
+
+- The invoke file editor now supports editing invocation files that use arrays or objects as contract parameters
+
 ## [2.1.49] - 2021-06-21
 
 ### Changed